+++ crates/topcoat-router/src/lib.rs
@@ -24,7 +24,10 @@ pub mod response;
 mod route;
 mod router;
 #[cfg(feature = "serve")]
+mod serve_dir;
+#[cfg(feature = "serve")]
 mod service;
+mod strip_prefix;
 #[cfg(feature = "tower")]
 pub mod tower;
 mod trailing_slash;
@@ -51,5 +54,8 @@ pub use query_param::*;
 pub use route::*;
 pub use router::*;
 #[cfg(feature = "serve")]
+pub use serve_dir::*;
+#[cfg(feature = "serve")]
 pub use service::*;
+pub use strip_prefix::*;
 pub use trailing_slash::*;